feat: Apply scroll to top
